Rust integration

+++ av1an/project/Project.py
@@ -6,10 +6,11 @@ from psutil import virtual_memory
 from distutils.spawn import find_executable
 from pathlib import Path
 from av1an.commandtypes import Command
-from av1an.utils import frame_probe_fast, hash_path, terminate
+from av1an.utils import frame_probe_fast
 from av1an.concat import vvc_concat, concatenate_ffmpeg, concatenate_mkvmerge
 from av1an.logger import log
 from av1an.vapoursynth import create_vs_file, frame_probe_vspipe
+from av1an.av1an import get_ffmpeg_info, hash_path
 
 
 class Project(object):
@@ -101,9 +102,12 @@ class Project(object):
         if self.frames > 0:
             return self.frames
 
-        if self.chunk_method in ('vs_ffms2', 'vs_lsmash'):
-            vs = self.input if self.is_vs else create_vs_file(
-                self.temp, self.input, self.chunk_method)
+        if self.chunk_method in ("vs_ffms2", "vs_lsmash"):
+            vs = (
+                self.input
+                if self.is_vs
+                else create_vs_file(self.temp, self.input, self.chunk_method)
+            )
             fr = frame_probe_vspipe(vs)
             if fr > 0:
                 self.frames = fr
@@ -123,17 +127,43 @@ class Project(object):
 
     def outputs_filenames(self):
         """
-        Set output filename
+        Set output filename and promts overwrite if file exists
 
         :param project: the Project
         """
         if self.webm:
-            suffix = '.webm'
+            suffix = ".webm"
         else:
-            suffix = '.mkv'
+            suffix = ".mkv"
 
-        self.output_file = Path(self.output_file).with_suffix(suffix) if self.output_file \
-            else Path(f'{self.input.stem}_{self.encoder}{suffix}')
+        # Check for non-empty string
+        if isinstance(self.output_file, str) and self.output_file:
+            if self.output_file[-1] in ("\\", "/"):
+                if not Path(self.output_file).exists():
+                    os.makedirs(Path(self.output_file), exist_ok=True)
+                self.output_file = Path(
+                    f"{self.output_file}{self.input.stem}_{self.encoder}{suffix}"
+                )
+            else:
+                self.output_file = Path(self.output_file).with_suffix(suffix)
+        else:
+            self.output_file = Path(f"{self.input.stem}_{self.encoder}{suffix}")
+
+    def promt_output_overwrite(self):
+
+        if self.output_file.exists():
+            print(
+                f":: Output file {self.output_file} exist, overwrite? [y/n or enter]:",
+                end="",
+            )
+
+            promt = input()
+
+            if "y" in promt.lower() or promt.strip() == "":
+                pass
+            else:
+                print("Stopping")
+                sys.exit()
 
     def load_project_from_file(self, path_string):
         """
@@ -149,7 +179,7 @@ class Project(object):
         Save project attributes from json to this class
         """
         pth = Path(path_string)
-        with open(pth, 'w') as json_data:
+        with open(pth, "w") as json_data:
             json_data.write(self.save_project())
 
     def save_project(self):
@@ -157,11 +187,11 @@ class Project(object):
         Returns json of this class, which later can be loaded
         """
         dt = dict(self.__dict__)
-        del dt['input']
-        del dt['output_file']
-        del dt['temp']
-        del dt['vmaf_path']
-        del dt['config']
+        del dt["input"]
+        del dt["output_file"]
+        del dt["temp"]
+        del dt["vmaf_path"]
+        del dt["config"]
         return json.dumps(dt, indent=4, sort_keys=True)
 
     def determine_workers(self):
@@ -170,15 +200,15 @@ class Project(object):
             return self.workers
 
         cpu = os.cpu_count()
-        ram = round(virtual_memory().total / 2**30)
+        ram = round(virtual_memory().total / 2 ** 30)
 
-        if self.encoder in ('aom', 'rav1e', 'vpx'):
+        if self.encoder in ("aom", "rav1e", "vpx"):
             workers = round(min(cpu / 3, ram / 1.5))
 
-        elif self.encoder in ('svt_av1', 'svt_vp9', 'x265', 'x264'):
+        elif self.encoder in ("svt_av1", "svt_vp9", "x265", "x264"):
             workers = round(min(cpu, ram)) // 8
 
-        elif self.encoder in 'vvc':
+        elif self.encoder in "vvc":
             workers = round(min(cpu, ram)) // 4
 
         # fix if workers round up to 0
@@ -190,21 +220,26 @@ class Project(object):
     def setup(self):
         """Creating temporally folders when needed."""
 
-        if self.temp:
-            self.temp = Path(str(self.temp))
-        else:
-            self.temp = Path('.' + str(hash_path(str(self.input))))
+        hash = str(hash_path(str(self.input)))
 
+        if self.temp:
+            if self.temp[-1] in ("\\", "/"):
+                self.temp = Path(f"{self.temp}{'.' + hash}")
+            else:
+                self.temp = Path(str(self.temp))
+        else:
+            self.temp = Path("." + hash)
+
+        log(f"File hash: {hash}")
         # Checking is resume possible
-        done_path = self.temp / 'done.json'
+        done_path = self.temp / "done.json"
         self.resume = self.resume and done_path.exists()
 
-        if not self.resume:
-            if self.temp.is_dir():
-                shutil.rmtree(self.temp)
+        if not self.resume and self.temp.is_dir():
+            shutil.rmtree(self.temp)
 
-        (self.temp / 'split').mkdir(parents=True, exist_ok=True)
-        (self.temp / 'encode').mkdir(exist_ok=True)
+        (self.temp / "split").mkdir(parents=True, exist_ok=True)
+        (self.temp / "encode").mkdir(exist_ok=True)
 
     def concat_routine(self):
         """
@@ -214,8 +249,8 @@ class Project(object):
         :return: None
         """
         try:
-            if self.encoder == 'vvc':
-                vvc_concat(self.temp, self.output_file.with_suffix('.h266'))
+            if self.encoder == "vvc":
+                vvc_concat(self.temp, self.output_file.with_suffix(".h266"))
             elif self.mkvmerge:
                 concatenate_mkvmerge(self.temp, self.output_file)
             else:
@@ -223,61 +258,76 @@ class Project(object):
         except Exception as e:
             _, _, exc_tb = sys.exc_info()
             print(
-                f'Concatenation failed, error At line: {exc_tb.tb_lineno}\nError:{str(e)}'
+                f"Concatenation failed, error At line: {exc_tb.tb_lineno}\nError:{str(e)}"
             )
-            log(f'Concatenation failed, aborting, error: {e}')
-            terminate()
+            log(f"Concatenation failed, aborting, error: {e}")
+            sys.exit(1)
 
     def select_best_chunking_method(self):
         """
         Selecting best chunking method based on available methods
         """
-        if not find_executable('vspipe'):
-            self.chunk_method = 'hybrid'
-            log('Set Chunking Method: Hybrid')
+        if not find_executable("vspipe"):
+            self.chunk_method = "hybrid"
+            log("Set Chunking Method: Hybrid")
         else:
             try:
                 import vapoursynth
+
                 plugins = vapoursynth.get_core().get_plugins()
 
-                if 'systems.innocent.lsmas' in plugins:
-                    log('Set Chunking Method: L-SMASH')
-                    self.chunk_method = 'vs_lsmash'
+                if "systems.innocent.lsmas" in plugins:
+                    log("Set Chunking Method: L-SMASH")
+                    self.chunk_method = "vs_lsmash"
 
-                elif 'com.vapoursynth.ffms2' in plugins:
-                    log('Set Chunking Method: FFMS2')
-                    self.chunk_method = 'vs_ffms2'
+                elif "com.vapoursynth.ffms2" in plugins:
+                    log("Set Chunking Method: FFMS2")
+                    self.chunk_method = "vs_ffms2"
+                else:
+                    log(f"Vapoursynth installed but no supported chunking methods.")
+                    log("Fallback to Hybrid")
+                    self.chunk_method = "hybrid"
 
             except Exception as e:
-                log(f'Vapoursynth not installed but vspipe reachable')
-                log(f'Error:{e}' + 'Fallback to Hybrid')
-                self.chunk_method = 'hybrid'
+                log(f"Vapoursynth not installed but vspipe reachable")
+                log(f"Error:{e}" + "Fallback to Hybrid")
+                self.chunk_method = "hybrid"
 
     def check_exes(self):
         """
         Checking required executables
         """
 
-        if not find_executable('ffmpeg'):
-            print('No ffmpeg')
-            terminate()
+        if not find_executable("ffmpeg"):
+            print("No ffmpeg")
+            sys.exit(1)
+        else:
+            log("Rust code")
+            log(get_ffmpeg_info())
 
-        if self.chunk_method in ['vs_ffms2', 'vs_lsmash']:
-            if not find_executable('vspipe'):
-                print('vspipe executable not found')
-                terminate()
+        if self.chunk_method in ["vs_ffms2", "vs_lsmash"]:
+            if not find_executable("vspipe"):
+                print("vspipe executable not found")
+                sys.exit(1)
 
             try:
                 import vapoursynth
+
                 plugins = vapoursynth.get_core().get_plugins()
 
-                if self.chunk_method == 'vs_lsmash' and "systems.innocent.lsmas" not in plugins:
-                    print('lsmas is not installed')
-                    terminate()
+                if (
+                    self.chunk_method == "vs_lsmash"
+                    and "systems.innocent.lsmas" not in plugins
+                ):
+                    print("lsmas is not installed")
+                    sys.exit(1)
 
-                if self.chunk_method == 'vs_ffms2' and "com.vapoursynth.ffms2" not in plugins:
-                    print('ffms2 is not installed')
-                    terminate()
+                if (
+                    self.chunk_method == "vs_ffms2"
+                    and "com.vapoursynth.ffms2" not in plugins
+                ):
+                    print("ffms2 is not installed")
+                    sys.exit(1)
             except ModuleNotFoundError:
-                print('Vapoursynth is not installed')
-                terminate()
+                print("Vapoursynth is not installed")
+                sys.exit(1)
